feat: 评分角色列表选择，并根据所选的角色筛选评分考核任务，评分和最终评分结果都加到被考核单位总分上。修改总分类型为Double。 (#153)
* fix: 考核内容校对分数总分没有归零会持续增加

* feat: 评分后自动修改相关考核内容节点的最高分和最低分

* fix: 如果考核任务未发布修改被考核单位的信息会报错

* feat: 评分后更新考核要点的各项得分

评分后更新考核要点的最高得分，最低得分，平均得分，最终得分。其中默认最高得分作为最终得分

* feat: 评分被考核单位添加权限控制

评分人只能看到自己所在考核组或者考核单位所负责单位的答题信息

* refactor: 评分增加两个字段

评分增加评分人所在考核组或者考核单位，以及相应类别字段

* fix: 多个被考核单位签收后评分时更新上级分数出错

评分时更新上级分数增加一个查询条件，只更新同一个mainId下的上级考核节点分数

* fix: 修复最终评分，未签收的被考核单位不显示的问题

* feat: 最终评分页面要点最终得分方式调整

* fix: 最终得分调整时，没有被考核单位时异常

* feat: 评分和最终评分结果都加到被考核单位总分上。修改总分类型为Double。

* feat: 最终评分页面增加被考核单位的总分字段

* feat: 评分角色列表选择，并根据所选的角色筛选评分考核任务

* fix: 排名中与答题信息表总分有关的数据修改为Double类型

+++ smart-system-core/src/main/java/org/jeecg/modules/smartAnswerRank/vo/Rank.java
@@ -43,13 +43,13 @@ public class Rank implements Serializable {
 
     private Map<String, Double> scoreMap;
 
-    private Map<String, Integer> columnMap;
+    private Map<String, Double> columnMap;
 
 
     /**总分*/
     @Excel(name = "总分", width = 15)
     @ApiModelProperty(value = "总分")
-    private Integer totalScore;
+    private Double totalScore;
 	/**排名*/
 	@Excel(name = "排名", width = 15)
     @ApiModelProperty(value = "排名")
